requirements.txt
flask==2.3.3
gunicorn==21.2.0
requests==2.31.0
beautifulsoup4==4.12.2
pdfminer.six==20221105
nltk==3.8.1
pymongo==4.6.0
schedule==1.2.1
twilio==8.5.0
python-dotenv==1.0.0
